refactor: QuizClearType 파일 common 모듈 enums 패키지로 이동
